Move country formatting back into base helper.
