+++ backend/app/controllers/spree/admin/base_controller.rb
@@ -14,7 +14,6 @@ module Spree
       helper 'spree/admin/navigation'
       helper 'spree/locale'
       helper 'spree/currency'
-      helper 'spree/country'
       layout 'spree/layouts/admin'
 
       before_action :authorize_admin
