+++ frontend/app/controllers/spree/store_controller.rb
@@ -15,7 +15,6 @@ module Spree
     helper 'spree/base'
     helper 'spree/locale'
     helper 'spree/currency'
-    helper 'spree/country'
 
     skip_before_action :verify_authenticity_token, only: :ensure_cart, raise: false
